chore: extracted provider data structures to another file
